Move SimpleList to a dedicated folder
