+++ packages/ra-ui-materialui/src/list/SimpleList/index.ts
@@ -1,0 +1,3 @@
+export * from './SimpleList';
+export * from './SimpleListLoading';
+export * from './SimpleListConfigurable';
