+++ packages/ra-ui-materialui/src/list/index.ts
@@ -1,5 +1,6 @@
 export * from './BulkActionsToolbar';
 export * from './datagrid';
+export * from './SimpleList';
 export * from './Empty';
 export * from './filter';
 export * from './FilterContext';
@@ -11,6 +12,4 @@ export * from './listFieldTypes';
 export * from './ListToolbar';
 export * from './pagination';
 export * from './Placeholder';
-export * from './SimpleList';
-export * from './SimpleListLoading';
 export * from './SingleFieldList';
